requirements.txt
fastapi==0.103.1
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
uvicorn==0.23.2
redis==5.0.2
